+++ conformance/runtime/connect_rpc/conformance/endpoint.ex
@@ -1,0 +1,13 @@
+defmodule ConnectRPC.Conformance.Endpoint do
+  @moduledoc false
+
+  use Plug.Builder
+
+  plug(Plug.Parsers,
+    parsers: [:urlencoded, :multipart, ConnectRPC.Parser, :json],
+    pass: ["*/*"],
+    json_decoder: Jason
+  )
+
+  plug(ConnectRPC.Conformance.Router)
+end